изменен формат получения видео и фото, добавлено обнаружение движение на камере
* fix cameras

* created ocv service

* remake logs

+++ microservices/users/internal/service/transport.go
@@ -7,11 +7,13 @@ import (
 	"io"
 	"net/http"
 
+	"github.com/Impisigmatus/service_core/log"
 	"github.com/Impisigmatus/service_core/utils"
 	"github.com/LeonKote/PSSVTelegramBot/microservices/users/autogen/server"
 	"github.com/LeonKote/PSSVTelegramBot/microservices/users/internal/models"
 	"github.com/LeonKote/PSSVTelegramBot/microservices/users/internal/repository"
 	jsoniter "github.com/json-iterator/go"
+	"github.com/rs/zerolog"
 )
 
 type Transport struct {
@@ -40,24 +42,30 @@ func NewTransport(repo repository.IUsersRepository) server.ServerInterface {
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) PostApiUsersAdd(w http.ResponseWriter, r *http.Request) {
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
+		return
+	}
+
 	data, err := io.ReadAll(r.Body)
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
 		return
 	}
 
 	var user models.User
 	if err := jsoniter.Unmarshal(data, &user); err != nil {
-		utils.WriteString(w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
+		utils.WriteString(log, w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
 		return
 	}
 
 	if err := transport.repo.AddUser(user); err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось добавить пользователя")
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось добавить пользователя")
 		return
 	}
 
-	utils.WriteNoContent(w)
+	utils.WriteNoContent(log, w)
 }
 
 // Set godoc
@@ -75,17 +83,23 @@ func (transport *Transport) PostApiUsersAdd(w http.ResponseWriter, r *http.Reque
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) DeleteApiUsersDeleteChatId(w http.ResponseWriter, r *http.Request, chatId int) {
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
+		return
+	}
+
 	ok, err := transport.repo.RemoveUser(int64(chatId))
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, err, "Пользователя не существует")
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Пользователя не существует")
 		return
 	}
 
 	if ok {
-		utils.WriteString(w, http.StatusOK, nil, "Пользователя не существует")
+		utils.WriteString(log, w, http.StatusOK, nil, "Пользователя не существует")
 		return
 	} else {
-		utils.WriteNoContent(w)
+		utils.WriteNoContent(log, w)
 		return
 	}
 }
@@ -105,18 +119,24 @@ func (transport *Transport) DeleteApiUsersDeleteChatId(w http.ResponseWriter, r 
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) GetApiUsersGetChatId(w http.ResponseWriter, r *http.Request, chatId int) {
-	user, err := transport.repo.GetUserByChatID(int64(chatId))
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			utils.WriteNoContent(w)
-			return
-		}
-
-		utils.WriteString(w, http.StatusNoContent, err, "Не удалось получить пользователя")
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
 		return
 	}
 
-	utils.WriteObject(w, user)
+	user, err := transport.repo.GetUserByChatID(int64(chatId))
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			utils.WriteNoContent(log, w)
+			return
+		}
+
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось получить пользователя")
+		return
+	}
+
+	utils.WriteObject(log, w, user)
 }
 
 // Set godoc
@@ -133,17 +153,23 @@ func (transport *Transport) GetApiUsersGetChatId(w http.ResponseWriter, r *http.
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) GetApiUsersGet(w http.ResponseWriter, r *http.Request) {
-	users, err := transport.repo.GetAllUsers()
-	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось получить пользователей")
-		return
-	}
-	if len(users) == 0 {
-		utils.WriteString(w, http.StatusInternalServerError, err, "В базе нет пользователей")
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
 		return
 	}
 
-	utils.WriteObject(w, users)
+	users, err := transport.repo.GetAllUsers()
+	if err != nil {
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось получить пользователей")
+		return
+	}
+	if len(users) == 0 {
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "В базе нет пользователей")
+		return
+	}
+
+	utils.WriteObject(log, w, users)
 }
 
 // Set godoc
@@ -160,18 +186,24 @@ func (transport *Transport) GetApiUsersGet(w http.ResponseWriter, r *http.Reques
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) GetApiUsersGetAdmin(w http.ResponseWriter, r *http.Request) {
-	users, err := transport.repo.GetAdminUser()
-	if err != nil {
-		if users == (models.User{}) {
-			utils.WriteString(w, http.StatusOK, nil, "Пользователя не существует")
-			return
-		}
-
-		utils.WriteString(w, http.StatusNoContent, err, "Не удалось получить пользователя")
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
 		return
 	}
 
-	utils.WriteObject(w, users)
+	users, err := transport.repo.GetAdminUser()
+	if err != nil {
+		if users == (models.User{}) {
+			utils.WriteString(log, w, http.StatusOK, nil, "Пользователя не существует")
+			return
+		}
+
+		utils.WriteString(log, w, http.StatusNoContent, err, "Не удалось получить пользователя")
+		return
+	}
+
+	utils.WriteObject(log, w, users)
 }
 
 // Set godoc
@@ -190,30 +222,36 @@ func (transport *Transport) GetApiUsersGetAdmin(w http.ResponseWriter, r *http.R
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) PutApiUsersUpdate(w http.ResponseWriter, r *http.Request) {
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
+		return
+	}
+
 	data, err := io.ReadAll(r.Body)
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
 		return
 	}
 
 	var updateUser models.User
 	if err := jsoniter.Unmarshal(data, &updateUser); err != nil {
-		utils.WriteString(w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
+		utils.WriteString(log, w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
 		return
 	}
 
-	ok, err := transport.repo.UpdateUser(updateUser)
+	ok, err = transport.repo.UpdateUser(updateUser)
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось обновить данные пользователя")
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось обновить данные пользователя")
 		return
 	}
 
 	if ok {
-		utils.WriteNoContent(w)
+		utils.WriteNoContent(log, w)
 		return
 	} else {
 
-		utils.WriteString(w, http.StatusOK, nil, "Пользователя не существует")
+		utils.WriteString(log, w, http.StatusOK, nil, "Пользователя не существует")
 		return
 	}
 }
